requirements-dev.txt
# Development dependencies

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0        # Parallel test execution
pytest-timeout>=2.2.0       # Test timeout management

# Linting and validation
pyyaml>=6.0.1
yamllint>=1.33.0
jsonschema>=4.20.0          # SKILL.md schema validation

# Analytics (for local testing)
PyGithub>=2.1.1             # GitHub API client
pandas>=2.0.0               # Data analysis

# Optional: enhanced testing
# pytest-watch>=4.2.0       # Watch mode for tests
# pytest-html>=4.1.0        # HTML test reports
